Second Version
The contraction mapping works.
However one problem with the grid given by Mendoza is that for some
initial values of capital and asset, the c - (l**omega)/omega is
negative whatever the decision variables.
This implies that the utility function  cannot be computed for these
values. This affects the values of our value functions as they then
depend on which value we assigne to negative composite consumption (made
of leisure and good consumption).
I have to think on how to solve this issue.