Twelth version
I think this actually the right way to do.
As far as I understand, the mistake (in case he made on) by Mendoza
might be on page 803. Actually I think he programmed the thing write and
that all the paper makes sense.

However (while I think he did it right on the program) he wrote on the
paper:
rho_e = rho_n = rho = 4*Pi-1
and
rho_en = theta
However it is theta that rules the persistence of e and n. So it should
be :
rho_e = rho_n = rho = theta
and
rho_en = 4*Pi-1